Update README.md
A slider is a web component that displays a series of images or content in a slideshow format, allowing users to cycle through them by clicking on arrow buttons. The slider is commonly used on websites to showcase products, portfolios, or other visual content.

Slider-using-Javascript, created using HTML, CSS, and JavaScript, is a slider for a website. It likely includes several key features, such as:

A container element that holds the slider's content.
Navigation buttons, which allow users to move forward or backward in the slider.